requirements.txt
absl-py==2.1.0
albumentations==1.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
attrs==24.2.0
backcall==0.2.0
backports.lzma==0.0.14
beautifulsoup4==4.12.3
black==21.4b2
bleach==6.0.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.15.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==2.2.1
cycler==0.11.0
dataclasses==0.6
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
detectron2==0.5+cu102
Django==3.2.25
entrypoints==0.4
exceptiongroup==1.2.2
fastapi==0.103.2
fastjsonschema==2.21.1
fonttools==4.38.0
future==1.0.0
fvcore==0.1.5.post20221221
google-auth==2.37.0
google-auth-oauthlib==0.4.6
grpcio==1.62.3
h11==0.14.0
hydra-core==1.3.2
idna==3.10
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
iopath==0.1.8
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.3.2
jsonschema==4.17.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==3.0.2
mypy-extensions==1.0.0
nbclassic==1.1.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.6.0
networkx==2.6.3
notebook==6.5.7
notebook_shim==0.2.4
numpy==1.21.6
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.0
pandas==1.1.5
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.11.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==8.0.1
pkgutil_resolve_name==1.3.10
plotly==4.14.0
portalocker==2.7.0
prometheus-client==0.17.1
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==6.1.0
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydicom==2.4.4
pydot==2.0.0
Pygments==2.17.2
pynetdicom==2.0.2
pyparsing==3.1.4
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
python-multipart==0.0.8
pytz==2024.2
PyWavelets==1.3.0
PyYAML==6.0.1
pyzmq==26.2.0
qudida==0.0.4
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
retrying==1.3.4
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.4.1
sqlparse==0.4.4
starlette==0.27.0
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
toml==0.10.2
torch==1.7.0
torchvision==0.8.1
tornado==6.2
tqdm==4.67.1
traitlets==5.9.0
typed-ast==1.5.5
typing_extensions==4.7.1
urllib3==2.0.7
uvicorn==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
yacs==0.1.8
zipp==3.15.0
